+++ src/SmartClusters.ts
@@ -32,7 +32,7 @@ const enum Place {
 }
 
 const enum Manner {
-    Stop,
+    Plosive,
     Fricative,
     Nasal,
     Sibilant,
@@ -40,10 +40,15 @@ const enum Manner {
     // the data array items fit in 80 chars when indented.
     LatFric,
     LatAffric,
-    Affricate
+    Affricate,
+    Approx,
+    LatApprox,
+    Trill
 }
 
 class Segment {
+    static index: 'digraph' | 'ipa';
+
     ipa: string;
     digraph: string;
     voiced: boolean;
@@ -53,6 +58,24 @@ class Segment {
     constructor(arr: [string, string, boolean, Place, Manner]) {
         [this.ipa, this.digraph, this.voiced, this.place, this.manner] = arr;
     }
+
+    get isStop() {
+        return this.manner === Manner.Nasal || this.manner === Manner.Plosive;
+    }
+
+    get isPeripheral() {
+        return this.place === Place.Bilabial || this.place === Place.Velar;
+    }
+
+    get isApprox() {
+        return this.manner === Manner.Approx
+            || this.manner === Manner.LatApprox
+            || this.manner === Manner.Trill;
+    }
+
+    toString() {
+        return this[Segment.index];
+    }
 }
 
 class ClusterEngine {
@@ -60,19 +83,22 @@ class ClusterEngine {
      * Using a constructor from a tuple might be slower at runtime, but it only
      * runs once, and takes up fewer characters.
      */
-    private static segments: Segment[] = [
+    private static readonly segments = [
         // Bilabial, labio-dental
-        new Segment(['p',  'p',   false, Place.Bilabial,    Manner.Stop]),
-        new Segment(['b',  'b',   true,  Place.Bilabial,    Manner.Stop]),
+        new Segment(['p',  'p',   false, Place.Bilabial,    Manner.Plosive]),
+        new Segment(['b',  'b',   true,  Place.Bilabial,    Manner.Plosive]),
         new Segment(['ɸ',  'ph',  false, Place.Bilabial,    Manner.Fricative]),
         new Segment(['β',  'bh',  true,  Place.Bilabial,    Manner.Fricative]),
         new Segment(['f',  'f',   false, Place.Labiodental, Manner.Fricative]),
         new Segment(['v',  'v',   true,  Place.Labiodental, Manner.Fricative]),
         new Segment(['m',  'm',   true,  Place.Bilabial,    Manner.Nasal]),
         new Segment(['m',  'm',   true,  Place.Labiodental, Manner.Nasal]),
+        new Segment(['ʋ',  'vw',  true,  Place.Bilabial,    Manner.Approx]),
+        new Segment(['w',  'w',   true,  Place.Bilabial,    Manner.Approx]),
+        new Segment(['w',  'w',   true,  Place.Labiodental, Manner.Approx]),
         // Alveolar
-        new Segment(['t',  't',   false, Place.Alveolar,    Manner.Stop]),
-        new Segment(['d',  'd',   true,  Place.Alveolar,    Manner.Stop]),
+        new Segment(['t',  't',   false, Place.Alveolar,    Manner.Plosive]),
+        new Segment(['d',  'd',   true,  Place.Alveolar,    Manner.Plosive]),
         new Segment(['s',  's',   false, Place.Alveolar,    Manner.Sibilant]),
         new Segment(['z',  'z',   true,  Place.Alveolar,    Manner.Sibilant]),
         new Segment(['θ',  'th',  false, Place.Alveolar,    Manner.Fricative]),
@@ -88,17 +114,22 @@ class ClusterEngine {
         new Segment(['tʃ', 'ch',  false, Place.PostAlv,     Manner.Affricate]),
         new Segment(['dʒ', 'j',   true,  Place.PostAlv,     Manner.Affricate]),
         new Segment(['n',  'n',   true,  Place.Alveolar,    Manner.Nasal]),
+        new Segment(['ɹ',  'rh',  true,  Place.Alveolar,    Manner.Approx]),
+        new Segment(['l',  'l',   true,  Place.Alveolar,    Manner.LatApprox]),
+        new Segment(['r',  'r',   true,  Place.Alveolar,    Manner.Trill]),
         // Retroflex
-        new Segment(['ʈ',  'rt',  false, Place.Retroflex,   Manner.Stop]),
-        new Segment(['ɖ',  'rd',  true,  Place.Retroflex,   Manner.Stop]),
+        new Segment(['ʈ',  'rt',  false, Place.Retroflex,   Manner.Plosive]),
+        new Segment(['ɖ',  'rd',  true,  Place.Retroflex,   Manner.Plosive]),
         new Segment(['ʂ',  'sr',  false, Place.Retroflex,   Manner.Sibilant]),
         new Segment(['ʐ',  'zr',  true,  Place.Retroflex,   Manner.Sibilant]),
         new Segment(['ʈʂ', 'rts', false, Place.Retroflex,   Manner.Affricate]),
         new Segment(['ɖʐ', 'rdz', true,  Place.Retroflex,   Manner.Affricate]),
         new Segment(['ɳ',  'rn',  true,  Place.Retroflex,   Manner.Nasal]),
+        new Segment(['ɻ',  'rr',  true,  Place.Retroflex,   Manner.Approx]),
+        new Segment(['ɭ',  'rl',  true,  Place.Retroflex,   Manner.LatApprox]),
         // Palatal
-        new Segment(['c',  'ky',  false, Place.Palatal,     Manner.Stop]),
-        new Segment(['ɟ',  'gy',  true,  Place.Palatal,     Manner.Stop]),
+        new Segment(['c',  'ky',  false, Place.Palatal,     Manner.Plosive]),
+        new Segment(['ɟ',  'gy',  true,  Place.Palatal,     Manner.Plosive]),
         new Segment(['ɕ',  'sy',  false, Place.Palatal,     Manner.Sibilant]),
         new Segment(['ʑ',  'zy',  true,  Place.Palatal,     Manner.Sibilant]),
         new Segment(['ç',  'hy',  false, Place.Palatal,     Manner.Fricative]),
@@ -106,42 +137,42 @@ class ClusterEngine {
         new Segment(['tɕ', 'cy',  false, Place.Palatal,     Manner.Affricate]),
         new Segment(['dʑ', 'jy',  true,  Place.Palatal,     Manner.Affricate]),
         new Segment(['ɲ',  'ny',  true,  Place.Palatal,     Manner.Nasal]),
+        new Segment(['j',  'y',   true,  Place.Palatal,     Manner.Approx]),
         // Velar
-        new Segment(['k',  'k',   false, Place.Velar,       Manner.Stop]),
-        new Segment(['g',  'g',   true,  Place.Velar,       Manner.Stop]),
+        new Segment(['k',  'k',   false, Place.Velar,       Manner.Plosive]),
+        new Segment(['g',  'g',   true,  Place.Velar,       Manner.Plosive]),
         new Segment(['x',  'kh',  false, Place.Velar,       Manner.Fricative]),
         new Segment(['ɣ',  'gh',  true,  Place.Velar,       Manner.Fricative]),
         new Segment(['ŋ',  'ng',  true,  Place.Velar,       Manner.Nasal]),
+        new Segment(['ɰ',  'wy',  true,  Place.Velar,       Manner.Approx]),
         // Uvular
-        new Segment(['q',  'q',   false, Place.Uvular,      Manner.Stop]),
-        new Segment(['ɢ',  'gq',  true,  Place.Uvular,      Manner.Stop]),
+        new Segment(['q',  'q',   false, Place.Uvular,      Manner.Plosive]),
+        new Segment(['ɢ',  'gq',  true,  Place.Uvular,      Manner.Plosive]),
         new Segment(['χ',  'qh',  false, Place.Uvular,      Manner.Fricative]),
         new Segment(['ʁ',  'gqh', true,  Place.Uvular,      Manner.Fricative]),
         new Segment(['ɴ',  'nq',  true,  Place.Uvular,      Manner.Nasal])
     ];
 
-    private index: 'digraph' | 'ipa';
-
     constructor(public isIpa: boolean) {
-        this.index = isIpa ? 'ipa' : 'digraph';
+        Segment.index = isIpa ? 'ipa' : 'digraph';
     }
 
     applyAssimilations(word: string[]) {
         const nasalAssimilate = (ph1: string, ph2: string) => {
             const data1 = ClusterEngine.segments.find(el =>
-                el[this.index] === ph1);
+                el.toString() === ph1);
 
             if (data1 && data1.manner === Manner.Nasal) {
                 const data2 = ClusterEngine.segments.find(el =>
-                    el[this.index] === ph2);
+                    el.toString() === ph2);
 
-                if (data2) {
+                if (data2 && !data2.isApprox) {
                     const result = ClusterEngine.segments.find(el =>
                         el.place === data2!.place
                         && el.manner === Manner.Nasal);
 
                     if (result) {
-                        return result[this.index];
+                        return result.toString();
                     }
                 }
             }
@@ -151,11 +182,11 @@ class ClusterEngine {
 
         const voiceAssimilate = (ph1: string, ph2: string) => {
             const data2 = ClusterEngine.segments.find(el =>
-                el[this.index] === ph2);
+                el.toString() === ph2);
 
-            if (data2 && data2.manner !== Manner.Nasal) {
+            if (data2 && !data2.isApprox) {
                 const data1 = ClusterEngine.segments.find(el =>
-                    el[this.index] === ph1);
+                    el.toString() === ph1);
 
                 if (data1) {
                     const result = ClusterEngine.segments.find(el =>
@@ -164,7 +195,7 @@ class ClusterEngine {
                         && el.manner === data1!.manner);
 
                     if (result) {
-                        return result[this.index];
+                        return result.toString();
                     }
                 }
             }
@@ -188,16 +219,16 @@ class ClusterEngine {
             ph2: string
         ): [string, string] => {
             const data1 = ClusterEngine.segments.find(el =>
-                el[this.index] === ph1);
+                el.toString() === ph1);
 
             if (data1 && data1.place === Place.Alveolar) {
                 const data2 = ClusterEngine.segments.find(el =>
-                    el[this.index] === ph2);
+                    el.toString() === ph2);
 
                 if (
                     data2
-                    && [Place.Velar, Place.Bilabial].includes(data2.place)
-                    && [Manner.Stop, Manner.Nasal].includes(data2.manner)
+                    && data2.isPeripheral
+                    && data2.isStop
                     && data2.manner === data1.manner
                 ) {
                     return [ph2, ph1];
@@ -217,6 +248,45 @@ class ClusterEngine {
         }
 
         return newArr;
+    }
+
+    applyRejections(word: string[]) {
+        const rejectCluster = (ph1: string, ph2: string) => {
+            const data1 = ClusterEngine.segments.find(
+                el => el.toString() === ph1
+            );
+
+            if (data1 && data1.manner !== Manner.Sibilant) {
+                const data2 = ClusterEngine.segments.find(
+                    el => el.toString() === ph2
+                );
+
+                if (
+                    data2
+                    && data2.isApprox
+                    && data2.place === data1.place
+                ) {
+                    // Trills are allowed a bit more generously
+                    if (data2.manner === Manner.Trill) {
+                        return data1.isApprox
+                            || data1.manner === Manner.Nasal;
+                    }
+
+                    // Allow /ll/ etc
+                    return data1 !== data2;
+                }
+            }
+
+            return false;
+        };
+
+        for (let i = 0; i < word.length - 1; ++i) {
+            if (rejectCluster(word[i]!, word[i + 1]!)) {
+                return ['REJECT'];
+            }
+        }
+
+        return word;
     }
 }